UI fixes

+++ src/main/java/com/alisonnet/medicalsystem/employeeportal/webcontroller/HRController.java
@@ -28,27 +28,15 @@ public class HRController {
 
     @GetMapping("/approve-employee")
     public String getEmployeesToApprovePage(Model model){
-        // basic employees info & employees to approve
-        List<BasicEmployee> basicEmployees = basicEmployeeService.findAll();
 
-        //all approved employees
-        List<Employee> employees = employeeService.findAll();
-
-        // receive non approved employees
-        List<BasicEmployee> nonApprovedEmployees =
-                basicEmployees.stream()
-                        .filter(basicEmployee -> employees.stream()
-                                .noneMatch(employee -> employee.getBasicInfo().equals(basicEmployee))
-                        ).collect(Collectors.toList());
-
-        model.addAttribute("employeesToApprove", nonApprovedEmployees);
+        model.addAttribute("employeesToApprove", basicEmployeeService.getUnapprovedEmployees());
         return "hr-approve-requests";
     }
 
     @GetMapping("/approve-employee/{id}")
     public String getApprovePage(@PathVariable int id, Model model){
-        Optional<BasicEmployee> maybeBasicEmployee = basicEmployeeService.findById(id);
 
+        Optional<BasicEmployee> maybeBasicEmployee = basicEmployeeService.findById(id);
         if(maybeBasicEmployee.isEmpty())
             return "redirect:/employee-portal/hr/approve-employee";
 
@@ -83,6 +71,11 @@ public class HRController {
 
     @GetMapping("/approve-employee/reject/{id}")
     public String deleteBasicEmployee(@PathVariable int id){
+
+        Optional<BasicEmployee> maybeBasicEmployee = basicEmployeeService.findById(id);
+        if(maybeBasicEmployee.isEmpty())                                                    //Add exception
+            return "redirect:/employee-portal/hr/approve-employee";
+
         basicEmployeeService.deleteById(id);
         return "redirect:/employee-portal/hr/approve-employee";
     }
@@ -90,7 +83,6 @@ public class HRController {
 
     @PostMapping("/employee/save")
     public String handleSaveEmployeeChanges(@ModelAttribute Employee employee){
-
         basicEmployeeService.save(employee.getBasicInfo());
         employeeService.save(employee);
         return "redirect:/employee-portal/hr/employee";
@@ -107,7 +99,6 @@ public class HRController {
     public String getEmployeeEditProfilePageById(@PathVariable int id, Model model){
 
         Optional<Employee> maybeEmployee = employeeService.findById(id);
-
         if(maybeEmployee.isEmpty()){
             return "redirect:/employee-portal/hr/employee";
         }
@@ -124,7 +115,6 @@ public class HRController {
     public String createNewContract(@PathVariable int id, Model model){
 
         Optional<Employee> maybeEmployee = employeeService.findById(id);
-
         if(maybeEmployee.isEmpty()){
             return "redirect:/employee-portal/hr/employee";
         }
@@ -140,15 +130,15 @@ public class HRController {
                                                 @ModelAttribute Contract contract){
 
         Optional<Employee> maybeEmployee = employeeService.findById(id);
-
         if(maybeEmployee.isEmpty()){
             return "redirect:/employee-portal/hr/employee";
         }
+
         Employee employee = maybeEmployee.get();
         contract.setEmployee(employee);
         contractService.save(contract);
 
-        return "redirect:/employee-portal/hr/employee/" + employee.getId();
+        return "redirect:/employee-portal/hr/employee/" + id;
     }
